Fix release pipeline
In https://github.com/polygon-io/client-python/pull/395 we fixed a gitub action to support the poetry version that dependabot is using so we can parse the poetry.lock syntax. However, this broke the automated release pipeline to https://pypi.org/project/polygon-api-client/ since we needed to update it's ability to parse the new poetry.lock syntax too.