Fix release pipelin (#414)
In #395 and #413 we fixed a gitub action to support the poetry version that dependabot is using so we can parse the poetry.lock syntax. However, this broke the automated release pipeline to https://pypi.org/project/polygon-api-client/ since we needed to update it's ability to parse the new poetry.lock syntax too.

The impact here is that we have not released 1.8.x.